% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitseq_setPlot.R
\name{bitseq_setPlot}
\alias{bitseq_setPlot}
\title{Configuring the settings of the plots for the BitSeq example}
\usage{
bitseq_setPlot(plotName)
}
\arguments{
\item{plotName}{Name of the plot with either ".png"
or ".pdf" extension.}
}
\value{
Plotting frame with the specified file name and 
preset features.
}
\description{
Function for initializing the plotting frame with preset
features and with the specified file name.
}
\examples{
bitseq_setPlot("geneA_GPprofile.pdf")
bitseq_setPlot("geneA_GPprofile.png")

}
\author{
Hande Topa, \email{hande.topa@helsinki.fi}
}
\keyword{internal}
\keyword{plot}

