% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructTwoSampleModel.R
\name{constructTwoSampleModel}
\alias{constructTwoSampleModel}
\title{Constructing GP model with the specified kernels for two samples}
\usage{
constructTwoSampleModel(x, y, v, kernelTypes, params = NULL)
}
\arguments{
\item{x}{A list of one-column matrices which contain the input
values for each series, i.e., time points for GP models.
The values given in this vector are used in GP model, so if any
transformation is needed, remember to transform them before
constructing the model.}

\item{y}{A list of one-column matrices which contain the observed
values at the corresponding time points given in the \code{x} vector.}

\item{v}{A list of one-column matrices which contain the fixed
variances at the corresponding time points given in the \code{x} vector,
with the corresponding observations given in the \code{y} vector.}

\item{kernelTypes}{Character vector which contains the types of the
kernels which will be used in the GP models.
Kernel types: \code{'rbf'}, \code{'white'}, \code{'bias'}, \code{'fixedvariance'}.
Note that the lower bound for the length scale parameter of rbf kernel
is set to the minimum distance between consecutive time points in order
to mitigate potential overfitting problems.}

\item{params}{Values of the kernel parameters in their transformed form.
If not specified, default values are assigned.}
}
\value{
Return GP model constucted with the specified kernel settings.
}
\description{
Function for constructing the GP model with the specified kernels
(and parameters) for two samples.
}
\examples{
x=list(as.matrix(seq(1,10)), as.matrix(seq(1,10)))
y=list(as.matrix(sin(x[[1]])), as.matrix(cos(x[[2]])))
v=list(as.matrix(runif(10,0,0.2)), as.matrix(runif(10,0,0.2)))
kernelTypes=c("rbf","white","fixedvariance")
model2=constructTwoSampleModel(x,y,v,kernelTypes)
model1=constructModel(rbind(x[[1]], x[[2]]), rbind(y[[1]], y[[2]]), rbind(v[[1]], v[[2]]), kernelTypes)

}
\author{
Hande Topa, \email{hande.topa@helsinki.fi}; Antti Honkela, \email{antti.honkela@helsinki.fi}
}
\keyword{fixedvariance}
\keyword{model}

