% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedvarianceKernCompute.R
\name{fixedvarianceKernCompute}
\alias{fixedvarianceKernCompute}
\title{fixedvarianceKernCompute}
\usage{
fixedvarianceKernCompute(kern, x, x2)
}
\arguments{
\item{kern}{GP kernel structure which contains the
fixed variances to be included in the \code{fixedvariance}
kernel.}

\item{x}{One-column matrix which contains the input
values for the rows of the kernel.}

\item{x2}{One-column matrix which contains the input
values for the columns of the kernel. If missing,
a square matrix is created with \code{x} on both 
rows and columns.}
}
\value{
Return kernel with the fixed variances in the diagonal,
and zeros in the off-diagonal elements.
}
\description{
Function for computing the fixed variance kernel.
}
\author{
Hande Topa, \email{hande.topa@helsinki.fi}
}
\keyword{fixedvariance}
\keyword{internal}
