% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedvarianceKernDiagCompute.R
\name{fixedvarianceKernDiagCompute}
\alias{fixedvarianceKernDiagCompute}
\title{fixedvarianceKernDiagCompute}
\usage{
fixedvarianceKernDiagCompute(kern, x)
}
\arguments{
\item{kern}{GP kernel structure which contains the
fixed variances.}

\item{x}{One-column matrix which contains the input values
for the rows (or the columns) of the kernel.}
}
\value{
Return matrix which contains the diagonal elements of the
fixed variance kernel.
}
\description{
Function for computing the diagonal of the \code{fixedvariance}
kernel.
}
\author{
Hande Topa, \email{hande.topa@helsinki.fi}
}
\keyword{fixedvariance}
\keyword{internal}
