% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedvarianceKernExpandParam.R
\name{fixedvarianceKernExpandParam}
\alias{fixedvarianceKernExpandParam}
\title{fixedvarianceKernExpandParam}
\usage{
fixedvarianceKernExpandParam(kern, params)
}
\arguments{
\item{kern}{GP kernel structure which contains the
fixed variances.}

\item{params}{Parameter values with which the kernel will be
updated.}
}
\value{
Return kernel with the updated parameter values.
}
\description{
Function for expanding the parameters of the \code{fixedvariance} kernel.
As the parameters of the \code{fixedvariance} kernel are fixed, and thus
are not optimized, their values remain the same. This function is
written to comply with the \code{gptk} package.
}
\author{
Hande Topa, \email{hande.topa@helsinki.fi}
}
\keyword{fixedvariance}
\keyword{internal}
