% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedvarianceKernExtractParam.R
\name{fixedvarianceKernExtractParam}
\alias{fixedvarianceKernExtractParam}
\title{fixedvarianceKernExtractParam}
\usage{
fixedvarianceKernExtractParam(kern, only.values = TRUE,
  untransformed.values = TRUE)
}
\arguments{
\item{kern}{GP kernel structure which contains the
fixed variances.}

\item{only.values}{Default set to TRUE.}

\item{untransformed.values}{Default set to TRUE.}
}
\value{
Return parameters of the kernel.
}
\description{
Function for extracting the parameters of the \code{fixedvariance} kernel.
As the parameters of the \code{fixedvariance} kernel are fixed, the kernel
has no parameters to be optimized. Hence, this function returns a
NULL vector. This function is written to comply with the \code{gptk} package.
}
\author{
Hande Topa, \email{hande.topa@helsinki.fi}
}
\keyword{fixedvariance}
\keyword{internal}
