% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedvarianceKernGradient.R
\name{fixedvarianceKernGradient}
\alias{fixedvarianceKernGradient}
\title{fixedvarianceKernGradient}
\usage{
fixedvarianceKernGradient(kern, x, x2, covGrad)
}
\arguments{
\item{kern}{GP kernel structure which contains the
fixed variances.}

\item{x}{x}

\item{x2}{x2}

\item{covGrad}{covGrad}
}
\value{
Return value for the gradient of fixedvariance kernel.
}
\description{
Function for computing the gradient of the \code{fixedvariance} kernel.
As the parameters of the \code{fixedvariance} kernel are fixed, the 
gradient equals zero. This function is written to comply with
the \code{gptk} package.
}
\author{
Hande Topa, \email{hande.topa@helsinki.fi}
}
\keyword{fixedvariance}
\keyword{internal}
