% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedvarianceKernParamInit.R
\name{fixedvarianceKernParamInit}
\alias{fixedvarianceKernParamInit}
\title{fixedvarianceKernParamInit}
\usage{
fixedvarianceKernParamInit(kern)
}
\arguments{
\item{kern}{GP kernel structure.}
}
\value{
Return kernel with the fixed variances in its diagonal.
}
\description{
Function for initializing the parameters of the \code{fixedvariance}
kernel. The parameters of the \code{fixedvariance} kernel are
specified when constructing the model and always kept fixed.
}
\author{
Hande Topa, \email{hande.topa@helsinki.fi}
}
\keyword{fixedvariance}
\keyword{internal}
