% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGP.R
\name{plotTwoSampleGP}
\alias{plotTwoSampleGP}
\title{Plotting fitted GP models for two-sample model}
\usage{
plotTwoSampleGP(model, index, col_item = "gray", ylimits = NULL,
  write_xticks = TRUE, write_yticks = TRUE, jitterx = FALSE)
}
\arguments{
\item{model}{GP model to be plotted.}

\item{index}{sample index of the model to be plotted from the two-sample model.}

\item{col_item}{RGB color code which will be used for the color of the GP plot.}

\item{ylimits}{Numeric vector which contains minimum and maximum limits for the y axis.}

\item{write_xticks}{Boolean: whether to write x ticks and labels or not}

\item{write_yticks}{Boolean: whether to write y ticks and labels or not}

\item{jitterx}{Boolean: whether to jitter duplicated x values or not}
}
\value{
Creates the plot of the fitted GP model.
}
\description{
Function for plotting fitted GP models within its confidence region of 2 standard deviations
for the two-sample model.
}
\examples{
x=list(as.matrix(seq(1,10)), as.matrix(seq(1,10)))
y=list(as.matrix(sin(x[[1]])), as.matrix(cos(x[[2]])))
v=list(as.matrix(runif(10,0,0.2)), as.matrix(runif(10,0,0.2)))
kernelTypes=c("rbf","white","fixedvariance")
model=constructTwoSampleModel(x,y,v,kernelTypes)
col_item=getColorVector()[1]
index=1
ylimits=c(min(y[[index]])-0.1,max(y[[index]])+0.1)

plotTwoSampleGP(model,index,col_item,ylimits)

}
\author{
Hande Topa, \email{hande.topa@helsinki.fi}; Antti Honkela, \email{antti.honkela@helsinki.fi}
}
\keyword{plot}
\keyword{twoSample}
