% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqDATA-data.R
\docType{data}
\name{RNAseqDATA}
\alias{RNAseqDATA}
\title{Sample data obtained from example BitSeq output files}
\format{An object of class \code{list} of length 3.}
\usage{
data(RNAseqDATA)
}
\description{
Data from an RNA-seq time series experiment.
This data set contains mean and standard deviation information on the expression
levels of 12 transcripts (which were originated from 4 genes) at 10 time points
(0, 5, 10, 20, 40, 80, 160, 320, 640, 1280 mins) for three settings: 'gene',
'abstr' (absolute transcript), and 'reltr' (relative transcript) expression
levels. In addition, the fields 'gene_mapping' and 'time_mapping' includes
information which is useful to match the genes with transcripts and the time
points with data files, respectively.
}
\examples{
data(RNAseqDATA)
gpData=RNAseqDATA$gene
gpData=RNAseqDATA$abstr
gpData=RNAseqDATA$reltr
}
\keyword{datasets}

