% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpData-data.R
\docType{data}
\name{snpData}
\alias{snpData}
\title{Sample data for demonstrating the application of experimental evolution.}
\format{An object of class \code{list} of length 2.}
\usage{
data(snpData)
}
\description{
This sample data contains 5 replicates of counts and sequencing depth
information for 5 SNPs at the generations (0, 10, 20, 30, 40, 50, 60).
}
\examples{
data(snpData)
counts=snpData$counts
seq_depth=snpData$seq_depth
}
\keyword{datasets}

